refactor: place viewr types in own file
