+++ src/components/viewer/SlideActionsBar.tsx
@@ -1,6 +1,6 @@
 import { IconDeviceTv, IconRefresh } from "@tabler/icons-react";
 import type { ReactNode } from "react";
-import type { ActionButtonState } from "../../types/actions";
+import type { ActionButtonState } from "../../types/viewer";
 import { ActionButtonList } from "./ActionButtonList";
 
 export type SlideActionBarKey =
